Ignore unrecognized query parameters
Spectastic rejects unsolicited query parameters, but allows any for
operations that don't have parameters. Consider the `Null` operation:

    '/null': {
        'get': {
            'operationId': 'Null',
            'parameters': [],
            [...]
         }
    }

The HTTP request `GET /null?foo=bar` would be accepted. But if we change
`Null` to define a query parameter `baz` in `parameters` it would reject
`GET /null?foo=bar`. If we go the other direction and change spectastic
to reject all requests with query parameters in the `Null` case, it
would be inconsistent with how we validate HTTP headers ( unrecognized
headers are ignored).

Therefore we change spectastic to ignore unrecognized query parameters.
To summarize our reasoning: the swagger specfication is about validating
HTTP semantics, not application semantics. It's appealing to enforce
structure at the boundary of applications, but because HTTP is so
extensible we have to be lenient in spectastic, and defer stricter
enforcement downstream to the application.

The result is that header and query parameters follow the same
validation semantics.

Jira: PE-4761
Closes: #3